Doc Update + Grid Boundary Limiters
Added a statement to check that the space chosen to move a bug or player is within the grid params.

Added and updated documentation for multiple methods.

Added a method to move the player to a space directly with one argument as well as having the method that moves the player based on two arguments.