Bridges API Build Half way Done
Still need to translate InputHandler.

GameGrid/game_grid.xml was turned into a fragment and added statically to the main activity.

It looks the same and works the same but is now able to be generated and latered by multiple activity classes.

Still working on NamedSymbols enum and NamedColor enums. Switched most NamedColor variables to Color.colorName for testing.